pyproject.toml
[tool.poetry]
name = "online-store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
psycopg = "^3.2.3"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
celery = "^5.4.0"
redis = "^5.1.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.2.0"
daphne = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
